Move Role type to renku-model-tinytypes
