+++ renku-model-tiny-types/src/test/scala/io/renku/graph/model/RenkuTinyTypeGenerators.scala
@@ -23,6 +23,7 @@ import io.renku.generators.Generators
 import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators._
 import io.renku.graph.model.images.{ImageResourceId, ImageUri}
+import io.renku.graph.model.projects.Role
 import io.renku.graph.model.versions.{CliVersion, SchemaVersion}
 import io.renku.tinytypes.InstantTinyType
 import org.scalacheck.Gen
@@ -32,6 +33,9 @@ import java.time.{Instant, LocalDate, ZoneOffset}
 import scala.util.Random
 
 trait RenkuTinyTypeGenerators {
+
+  val roleGen: Gen[Role] =
+    Gen.oneOf(Role.all.toList)
 
   def associationResourceIdGen: Gen[associations.ResourceId] =
     Generators.validatedUrls.map(_.value).map(associations.ResourceId)
