+++ renku-model-tiny-types/src/test/scala/io/renku/graph/model/RoleSpec.scala
@@ -16,15 +16,15 @@
  * limitations under the License.
  */
 
-package io.renku.projectauth
+package io.renku.graph.model
 
 import cats.data.NonEmptyList
-import io.renku.projectauth.Role._
+import io.renku.graph.model.projects.Role
+import io.renku.graph.model.projects.Role._
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should
 
 class RoleSpec extends AnyFlatSpec with should.Matchers {
-
   it should "order roles correctly" in {
     val roles = Role.all.sorted
     roles                            shouldBe NonEmptyList.of(Reader, Maintainer, Owner)
